# api/openapi.yaml
openapi: 3.1.2
info:
  title: Synodus Core API
  version: 1.0.0
  description: >-
    The reviewed v1 browser API contract. Organization, team, and task
    operations remain dark until the D06 route cutover.
...
paths:
...
  /api/v1/orgs/{org_id}/resources:
    parameters: [{$ref: '#/components/parameters/OrganizationID'}]
    get:
      tags: [Resources]
      operationId: listResources
      security: [{sessionCookie: []}]
      responses:
        '200': {description: Resources, content: {application/json: {schema: {$ref: '#/components/schemas/ResourceCollection'}}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
    post:
      tags: [Resources]
      operationId: createResource
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ResourceWrite'}}}}
      responses:
        '201': {description: Resource created, content: {application/json: {schema: {$ref: '#/components/schemas/Resource'}}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/v1/orgs/{org_id}/resources/{resource_id}:
    parameters: [{$ref: '#/components/parameters/OrganizationID'}, {$ref: '#/components/parameters/ResourceID'}]
    patch:
      tags: [Resources]
      operationId: updateResource
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ResourcePatch'}}}}
      responses:
        '200': {description: Resource updated, content: {application/json: {schema: {$ref: '#/components/schemas/Resource'}}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
    delete:
      tags: [Resources]
      operationId: deleteResource
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses:
        '204': {description: Resource deleted}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/v1/orgs/{org_id}/bookings:
    parameters: [{$ref: '#/components/parameters/OrganizationID'}]
    get:
      tags: [Resources]
      operationId: listBookings
      security: [{sessionCookie: []}]
      responses:
        '200': {description: Approved bookings, content: {application/json: {schema: {$ref: '#/components/schemas/BookingCollection'}}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/v1/orgs/{org_id}/resources/{resource_id}/requests:
    parameters: [{$ref: '#/components/parameters/OrganizationID'}, {$ref: '#/components/parameters/ResourceID'}]
    post:
      tags: [Resources]
      operationId: createResourceRequest
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ResourceRequestWrite'}}}}
      responses:
        '201': {description: Request submitted, content: {application/json: {schema: {$ref: '#/components/schemas/ResourceRequest'}}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/v1/orgs/{org_id}/resource-requests:
    parameters: [{$ref: '#/components/parameters/OrganizationID'}]
    get:
      tags: [Resources]
      operationId: listResourceRequests
      security: [{sessionCookie: []}]
      responses:
        '200': {description: Visible resource requests, content: {application/json: {schema: {$ref: '#/components/schemas/ResourceRequestCollection'}}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/v1/orgs/{org_id}/resource-requests/{request_id}:
    parameters: [{$ref: '#/components/parameters/OrganizationID'}, {$ref: '#/components/parameters/RequestID'}]
    patch:
      tags: [Resources]
      operationId: decideResourceRequest
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ResourceDecision'}}}}
      responses:
        '200': {description: Request decided, content: {application/json: {schema: {$ref: '#/components/schemas/ResourceRequest'}}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-synodus_session
      description: Production browser session cookie.
...
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 512}
...
    OrganizationID:
      name: org_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/PublicID'}
...
    ResourceID:
      name: resource_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/PublicID'}
    RequestID:
      name: request_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/PublicID'}
  responses:
    InvalidRequest: {$ref: '#/components/responses/InvalidRequestResponse'}
    Unauthenticated: {$ref: '#/components/responses/UnauthenticatedResponse'}
    Forbidden: {$ref: '#/components/responses/ForbiddenResponse'}
    NotFound: {$ref: '#/components/responses/NotFoundResponse'}
    Conflict: {$ref: '#/components/responses/ConflictResponse'}
...
    InternalFailure: {$ref: '#/components/responses/InternalFailureResponse'}
    DependencyUnavailable: {$ref: '#/components/responses/DependencyUnavailableResponse'}
    InvalidRequestResponse:
      description: The request is invalid
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
          examples: {standard: {$ref: '#/components/examples/InvalidRequestProblem'}}
    UnauthenticatedResponse:
      description: Authentication is required
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
          examples: {standard: {$ref: '#/components/examples/UnauthenticatedProblem'}}
    ForbiddenResponse:
      description: The authenticated user cannot perform the known operation
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
          examples: {standard: {$ref: '#/components/examples/ForbiddenProblem'}}
    NotFoundResponse:
      description: The protected resource is absent or unavailable to the user
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
          examples: {standard: {$ref: '#/components/examples/NotFoundProblem'}}
    ConflictResponse:
      description: The request conflicts with current state
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
          examples: {standard: {$ref: '#/components/examples/ConflictProblem'}}
...
    InternalFailureResponse:
      description: An internal failure occurred
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
          examples: {standard: {$ref: '#/components/examples/InternalProblem'}}
    DependencyUnavailableResponse:
      description: A required dependency is unavailable
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
          examples: {standard: {$ref: '#/components/examples/DependencyProblem'}}
  schemas:
    PublicID: {type: string, format: uuid}
    Timestamp: {type: string, format: date-time}
...
    AvailabilityWindow:
      type: object
      additionalProperties: false
      required: [weekday, start, end]
      properties:
        weekday: {type: integer, minimum: 0, maximum: 6}
        start: {type: string, pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'}
        end: {type: string, pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'}
    ResourceWrite:
      type: object
      additionalProperties: false
      required: [name, code, kind, location, enabled, availability]
      properties:
        name: {type: string, minLength: 1, maxLength: 255}
        code: {type: string, minLength: 1, maxLength: 64}
        kind: {type: string, enum: [gpu, instrument, room, workstation]}
        location: {type: string, maxLength: 255}
        enabled: {type: boolean}
        availability: {type: array, maxItems: 7, items: {$ref: '#/components/schemas/AvailabilityWindow'}}
    ResourcePatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: {type: string, minLength: 1, maxLength: 255}
        code: {type: string, minLength: 1, maxLength: 64}
        kind: {type: string, enum: [gpu, instrument, room, workstation]}
        location: {type: string, maxLength: 255}
        enabled: {type: boolean}
        availability: {type: array, maxItems: 7, items: {$ref: '#/components/schemas/AvailabilityWindow'}}
    Resource:
      allOf:
        - {$ref: '#/components/schemas/ResourceWrite'}
        - {type: object, required: [id], properties: {id: {$ref: '#/components/schemas/PublicID'}}}
    Booking:
      type: object
      additionalProperties: false
      required: [id, resource_id, resource_name, start_time, end_time, details_visible, team_public_id, team_name, requested_by_name, purpose]
      properties:
        id: {$ref: '#/components/schemas/PublicID'}
        resource_id: {$ref: '#/components/schemas/PublicID'}
        resource_name: {type: string}
        start_time: {$ref: '#/components/schemas/Timestamp'}
        end_time: {$ref: '#/components/schemas/Timestamp'}
        details_visible: {type: boolean}
        team_public_id: {type: [string, 'null'], format: uuid}
        team_name: {type: [string, 'null']}
        requested_by_name: {type: [string, 'null']}
        purpose: {type: [string, 'null']}
    ResourceRequestWrite:
      type: object
      additionalProperties: false
      required: [team_public_id, start_time, end_time, purpose]
      properties:
        team_public_id: {$ref: '#/components/schemas/PublicID'}
        start_time: {$ref: '#/components/schemas/Timestamp'}
        end_time: {$ref: '#/components/schemas/Timestamp'}
        purpose: {type: string, minLength: 1, maxLength: 1000}
    ResourceDecision:
      type: object
      additionalProperties: false
      required: [status]
      properties: {status: {type: string, enum: [approved, rejected]}}
    ResourceRequest:
      type: object
      additionalProperties: false
      required: [id, resource_id, resource_name, team_public_id, team_name, requested_by, requested_by_name, start_time, end_time, purpose, status, created_at, decided_at]
      properties:
        id: {$ref: '#/components/schemas/PublicID'}
        resource_id: {$ref: '#/components/schemas/PublicID'}
        resource_name: {type: string}
        team_public_id: {$ref: '#/components/schemas/PublicID'}
        team_name: {type: string}
        requested_by: {$ref: '#/components/schemas/PublicID'}
        requested_by_name: {type: string}
        start_time: {$ref: '#/components/schemas/Timestamp'}
        end_time: {$ref: '#/components/schemas/Timestamp'}
        purpose: {type: string}
        status: {type: string, enum: [pending, approved, rejected]}
        created_at: {$ref: '#/components/schemas/Timestamp'}
        decided_at: {type: [string, 'null'], format: date-time}
    ResourceCollection: {type: object, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/Resource'}}}}
    BookingCollection: {type: object, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/Booking'}}}}
    ResourceRequestCollection: {type: object, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/ResourceRequest'}}}}
...
    Violation:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field: {type: string, minLength: 1, maxLength: 64}
        code: {type: string, minLength: 1, maxLength: 64}
        message: {type: string, minLength: 1, maxLength: 256}
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, request_id]
      properties:
        type:
          type: string
          enum: [/problems/invalid-request, /problems/unauthenticated, /problems/forbidden, /problems/not-found, /problems/conflict, /problems/precondition-failed, /problems/rate-limited, /problems/internal-failure, /problems/dependency-unavailable]
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string, minLength: 1, maxLength: 512}
        request_id: {type: string, format: uuid}
        violations:
          type: array
          maxItems: 16
          items: {$ref: '#/components/schemas/Violation'}
  examples:
...
    InvalidRequestProblem:
      value: {type: /problems/invalid-request, title: Invalid request, status: 400, detail: The request is invalid., request_id: 55555555-5555-4555-8555-555555555555, violations: [{field: body, code: malformed, message: The request body must contain one valid JSON object.}]}
    UnauthenticatedProblem:
      value: {type: /problems/unauthenticated, title: Authentication required, status: 401, detail: Authentication is required., request_id: 55555555-5555-4555-8555-555555555555}
    ForbiddenProblem:
      value: {type: /problems/forbidden, title: Forbidden, status: 403, detail: The operation is not permitted., request_id: 55555555-5555-4555-8555-555555555555}
    NotFoundProblem:
      value: {type: /problems/not-found, title: Not found, status: 404, detail: The requested resource was not found., request_id: 55555555-5555-4555-8555-555555555555}
    ConflictProblem:
      value: {type: /problems/conflict, title: Conflict, status: 409, detail: The request conflicts with current state., request_id: 55555555-5555-4555-8555-555555555555}
...
    InternalProblem:
      value: {type: /problems/internal-failure, title: Internal server error, status: 500, detail: An unexpected error occurred., request_id: 55555555-5555-4555-8555-555555555555}
    DependencyProblem:
      value: {type: /problems/dependency-unavailable, title: Dependency unavailable, status: 503, detail: A required dependency is unavailable., request_id: 55555555-5555-4555-8555-555555555555}