# internal/repository/tenantdb/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package tenantdb

import (
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Resource struct {
	ID           int64     `json:"id"`
	PublicID     uuid.UUID `json:"public_id"`
	Name         string    `json:"name"`
	Code         string    `json:"code"`
	Kind         string    `json:"kind"`
	Location     string    `json:"location"`
	Enabled      bool      `json:"enabled"`
	Availability []byte    `json:"availability"`
	CreatedAt    time.Time `json:"created_at"`
	UpdatedAt    time.Time `json:"updated_at"`
}

type ResourceRequest struct {
	ID          int64              `json:"id"`
	PublicID    uuid.UUID          `json:"public_id"`
	ResourceID  int64              `json:"resource_id"`
	TeamID      int64              `json:"team_id"`
	RequestedBy uuid.UUID          `json:"requested_by"`
	StartTime   time.Time          `json:"start_time"`
	EndTime     time.Time          `json:"end_time"`
	Purpose     string             `json:"purpose"`
	Status      string             `json:"status"`
	CreatedAt   time.Time          `json:"created_at"`
	DecidedAt   pgtype.Timestamptz `json:"decided_at"`
}

type Task struct {
	ID          int64     `json:"id"`
	TeamID      int64     `json:"team_id"`
	Description string    `json:"description"`
	Status      string    `json:"status"`
	CreatedAt   time.Time `json:"created_at"`
	UpdatedAt   time.Time `json:"updated_at"`
	PublicID    uuid.UUID `json:"public_id"`
}

type Team struct {
	ID           int64     `json:"id"`
	Name         string    `json:"name"`
	Slug         string    `json:"slug"`
	CreatedAt    time.Time `json:"created_at"`
	UpdatedAt    time.Time `json:"updated_at"`
	PublicID     uuid.UUID `json:"public_id"`
	IsQuarantine bool      `json:"is_quarantine"`
}

type TeamMember struct {
	TeamID       int64     `json:"team_id"`
	UserPublicID uuid.UUID `json:"user_public_id"`
	Role         string    `json:"role"`
	CreatedAt    time.Time `json:"created_at"`
	UpdatedAt    time.Time `json:"updated_at"`
}